Further optimizations: calculateHogCells(...) kernel takes 9ms on a 640x480 image

+++ src/HogApp/hogprocessor.h
@@ -32,7 +32,7 @@ protected:
     void calculateHogOcl();
     void calculateHogPiotr();
     void compareDescriptors() const;
-    void compareDescriptorsOcl(const float *mappedDescriptor) const;
+    void compareDescriptorsOcl(const uint *mappedDescriptor) const;
 
     std::shared_ptr<cv::Mat_<uchar> > ocvImageGray_ = nullptr;
     std::shared_ptr<cv::Mat_<float> > ocvImageGrayFloat_ = nullptr;
